:constrcution:upgrade minikin from android11 to android14

+++ src/gui/text/minikin/include/minikin/LayoutCache.h
@@ -83,9 +83,9 @@ public:
 
 private:
     const uint16_t* mChars;
-    size_t mNchars;
-    size_t mStart;
-    size_t mCount;
+    uint32_t mNchars;
+    uint32_t mStart;
+    uint32_t mCount;
     uint32_t mId;  // for the font collection
     FontStyle mStyle;
     float mSize;
